// src/shared/dataprovider/api/model/activityLogDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface ActivityLogDTO { 
    readonly Id?: number;
    readonly UserName?: string;
    readonly ClientId?:number;
    readonly ClientName?: string;
    readonly RoleName?: string;
    readonly Action?: string;
    readonly Description?: string;
    readonly Timestamp?: Date;
}